Fix player stats loader from code review findings
- Rewrite etl/load_player_stats.py: sync httpx via shared client, fix
  missing Team import, parse CFBD's long-format /stats/player/season
  response, add CLI entrypoint matching load_rosters.py
- Centralize CFBD auth/base URL in etl/cfbd_client.py get() helper
  (Bearer header, 30s timeout, raise_for_status) and use it in both
  loaders
- Change sacks/tackles_for_loss to Float (CFBD reports half units)
- Add Alembic migration for player_stats_offense/player_stats_defense
- load_rosters --all: pass team_id down instead of re-resolving each
  school name per team per year

+++ db/models.py
@@ -1,5 +1,5 @@
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
-from sqlalchemy import String, Integer, BigInteger, ForeignKey
+from sqlalchemy import String, Integer, BigInteger, Float, ForeignKey
 
 class Base(DeclarativeBase): pass
 
@@ -25,12 +25,13 @@ class Roster(Base):
     position: Mapped[str | None] = mapped_column(String, nullable=True)
     jersey: Mapped[int | None] = mapped_column(nullable=True)
     player_cfbd_id: Mapped[int | None] = mapped_column(index=True, nullable=True)
+
 class PlayerStatsOffense(Base):
     __tablename__ = "player_stats_offense"
     season: Mapped[int] = mapped_column(Integer, primary_key=True)
     team_id: Mapped[int] = mapped_column(ForeignKey("teams.team_id"), primary_key=True)
     player_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
-    
+
     passing_yards: Mapped[int] = mapped_column(Integer)
     rushing_yards: Mapped[int] = mapped_column(Integer)
     receiving_yards: Mapped[int] = mapped_column(Integer)
@@ -45,9 +46,7 @@ class PlayerStatsDefense(Base):
     player_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
 
     tackles: Mapped[int] = mapped_column(Integer)
-    tackles_for_loss: Mapped[int] = mapped_column(Integer)
-    sacks: Mapped[int] = mapped_column(Integer)
+    tackles_for_loss: Mapped[float] = mapped_column(Float)
+    sacks: Mapped[float] = mapped_column(Float)
     interceptions: Mapped[int] = mapped_column(Integer)
     touchdowns: Mapped[int] = mapped_column(Integer)
-
-
